+++ src/features/carrito/components/index.js
@@ -1,0 +1,1 @@
+export { default as Pago } from "./Pago";
